DX-349: pull model names from $refs in swagger docs
This is to just get it working. From here would
be ideal to extract most of the new HTML out of
endpoint.jsx and into a separate React component

Reason for passing $refs in this way was because
when dereferencing swagger docs, it gets rid of
$ref value and theres no way of retrieving it
for the anchor in the api-reference page.
Therefore, we pass the docs twice, once dereferenced,
and once not so that we can refer to these $ref
tags. I could not find anything in Swagger docs
that supported something like this out of the box.