DX-343: make tests pass with new path structure
Optimized tests by not calling maximizeWindow and end
on every test, but only in before and after suite.

Added flag to test-web so that it calls npm run build
unless specified not to.

Modified all tests to use proper url paths (changed
in 342 ticket). Had to change a tad bit of logic in
tests but nothing major